[Issue#31] Initial pass at splitting up `addrman`
